workflows: Extract configuration
